Refactored and moved tests from harvester-api to datastore
