+++ libraries/datastore/src/test/java/no/dcat/datastore/domain/dcat/DcatConverterTest.java
@@ -1,20 +1,15 @@
-package no.dcat.harvester.crawler;
+package no.dcat.datastore.domain.dcat;
 
-import no.dcat.harvester.service.SubjectCrawler;
+import no.dcat.datastore.domain.dcat.builders.DcatReader;
 import no.dcat.shared.Contact;
 import no.dcat.shared.Dataset;
-import no.dcat.shared.Subject;
-import no.dcat.datastore.domain.dcat.builders.DcatReader;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.riot.RDFDataMgr;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 
@@ -22,21 +17,15 @@ import java.text.SimpleDateFormat;
 import java.util.List;
 
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertThat;
 
-@ActiveProfiles(value = "unit-integration")
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DcatSourcesConverterIT {
-    private static Logger logger = LoggerFactory.getLogger(DcatSourcesConverterIT.class);
+public class DcatConverterTest {
+    private static Logger logger = LoggerFactory.getLogger(DcatConverterTest.class);
     static SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
 
-    @Autowired
-    SubjectCrawler subjectCrawler;
 
     public DcatReader setupReader(Model model) {
-        return new DcatReader(model, "http://localhost:8100", "user", "password");
+        return new DcatReader(model);
     }
 
     @Test
@@ -51,22 +40,6 @@ public class DcatSourcesConverterIT {
 
     }
 
-    @Test
-    public void readDcatWithSubjectReference() throws Throwable {
-        Resource r = new ClassPathResource("begrepHarvest.ttl");
-        Model model = new CrawlerJob(null,null,null,subjectCrawler).loadModelAndValidate(r.getURL());
-
-        model.write(System.out, "TURTLE");
-
-        DcatReader reader = setupReader(model);
-        List<Subject> actualSubjects = reader.getSubjects();
-
-        assertThat(actualSubjects.size(), is(1));
-        Subject actualSubject = actualSubjects.get(0);
-
-        assertThat(actualSubject.getPrefLabel().get("no"), is("enhet") );
-        assertThat(actualSubject.getCreator().getUri(), is("http://data.brreg.no/enhetsregisteret/enhet/974760673"));
-    }
 
     @Test
     public void readRamsundContactInfoOK() throws Throwable {
@@ -116,30 +89,6 @@ public class DcatSourcesConverterIT {
         assertThat("has temporals", t[0], is(10));
     }
 
-    @Test
-    public void readCompleteDifiData() throws Throwable {
-
-        Resource r = new ClassPathResource("difi-complete-2017-10-25.jsonld");
-        Model model = new CrawlerJob(null,null,null, subjectCrawler).loadModelAndValidate(r.getURL());
-
-        DcatReader reader = setupReader(model);
-        List<Dataset> datasets = reader.getDatasets();
-
-        datasets.forEach(dataset -> {
-            if (dataset.getDistribution() != null) {
-                dataset.getDistribution().forEach( distribution -> {
-                    logger.debug("{}: {}", dataset.getUri(), distribution.getFormat());
-                    if (distribution.getFormat() != null && distribution.getFormat().size() > 0) {
-                        String firstFormat = distribution.getFormat().get(0);
-                        assertThat(firstFormat, is(notNullValue()));
-                    }
-
-                });
-            }
-        });
-
-        assertThat(datasets.size() , is(484));
-    }
 
     @Test
     public void readGeonorgeData() throws Throwable {
