+++ libraries/datastore/src/test/resources/gdoc-data-2017-10-19.ttl
