+++ libraries/datastore/src/test/resources/geonorge-data-2017-10-19.xml
