+++ libraries/datastore/src/test/resources/nav-2017-10-31.ttl
