+++ libraries/datastore/src/test/resources/reinli-data-2017-10-26.ttl
